refactor: remove unnecessary converting & change compare logic
